Memories as Spins
Added functionality to track the long-range correlations of changes in the memory DOFs. The sign of the derivative of a memory DOF is treated as its "spin" (up or down), and its correlations are then detected by extracting avalanches.

NOTE: At the moment, due to the higher connectivity of the "memories as spins" in the lattice (6 nearest neighbors instead of 4), avalanche extraction scales extremely poorly with system size.